resen problem z sejami pri urejanju vprasanja

+++ Vprasanja/PSiA/delete.php
@@ -1,7 +1,11 @@
 <?php
 include("config.php");
 session_start();
-$id = $_SESSION['session_id'];
+if (!isset($_SESSION['logged_on'])) {
+  header("location: index.php");
+  die();
+}
+$id = $_GET['id'];
 mysqli_query($db, "DELETE FROM data WHERE ID='$id'");
 
 header("location:admin.php");
